+++ examples/experimental/i3s-picking/package.json
@@ -9,19 +9,19 @@
     "generate": "node ./generate-index/index.js"
   },
   "dependencies": {
-    "@deck.gl/core": "8.4.0-beta.2",
-    "@deck.gl/geo-layers": "8.4.0-beta.2",
-    "@deck.gl/layers": "8.4.0-beta.2",
-    "@deck.gl/mesh-layers": "8.4.0-beta.2",
-    "@deck.gl/react": "8.4.0-beta.2",
+    "@deck.gl/core": "8.4.6",
+    "@deck.gl/geo-layers": "8.4.6",
+    "@deck.gl/layers": "8.4.6",
+    "@deck.gl/mesh-layers": "8.4.6",
+    "@deck.gl/react": "8.4.6",
     "@loaders.gl/core": "2.2.0",
     "@loaders.gl/draco": "2.2.0",
     "@loaders.gl/gltf": "2.2.0",
     "@loaders.gl/i3s": "2.2.0",
-    "@luma.gl/constants": "^8.4.0-beta.2",
-    "@luma.gl/core": "^8.4.0-beta.2",
-    "@math.gl/core": "^3.3.0",
-    "@math.gl/geospatial": "^3.3.0",
+    "@luma.gl/constants": "^8.4.3",
+    "@luma.gl/core": "^8.4.3",
+    "@math.gl/core": "^3.4.0",
+    "@math.gl/geospatial": "^3.4.0",
     "@probe.gl/stats-widget": "^3.3.0",
     "probe.gl": "^3.3.0",
     "prop-types": "^15.7.2",
